# [Xamarin.Android.Build.Tasks] leave classes.zip uncompressed