@jonathanpeppers 2018-09-04
Build logs here: [ClassesZipUncompressed.zip](https://github.com/xamarin/xamarin-android/files/2349923/ClassesZipUncompressed.zip)